requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.0
aiohttp~=3.9.5
APScheduler==3.11.2
SQLAlchemy==2.0.25
alembic==1.13.1
aiosqlite==0.20.0
asyncpg==0.30.0
